requirements.txt
africastalking==1.2.4
asgiref==3.4.1
beautifulsoup4==4.10.0
bootstrap4==0.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
cloudinary==1.26.0
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-admin-black @ git+https://github.com/app-generator/django-admin-black.git@fb2da2b86022a591b217bd9ca665e64b9f0abf9c
django-bookmark-base==2020.7.1
django-bootstrap4==3.0.1
django-braces==1.14.0
django-crispy-forms==1.13.0
django-model-utils==4.1.1
django-mpesa==2.0.8
django-social-share==2.2.1
django-star-ratings==0.9.1
djangorestframework==3.12.4
drf-yasg==1.20.0
Faker==9.2.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
Markdown==3.3.4
MarkupSafe==2.0.1
mpesa==0.0.1
mpesa-api==0.2.3
packaging==21.0
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
schema==0.7.4
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
swapper==1.1.2.post1
text-unidecode==1.3
toposort==1.7
uritemplate==4.0.0
urllib3==1.26.7
